Add TODOs for missing execution paths
